add navigation utility
