+++ src/components/Search/NavigationItem/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { SearchNavigationResult, SearchNavigationType } from 'types/SearchNavigationResult';
+import { SearchNavigationResult } from 'types/SearchNavigationResult';
 import Link from 'next/link';
-import { getVerseNavigationUrl } from 'src/utils/verse';
+import { resolveUrlBySearchNavigationType } from 'src/utils/navigation';
 import styles from './NavigationItem.module.scss';
 
 interface Props {
@@ -9,7 +9,7 @@ interface Props {
 }
 
 const NavigationItem: React.FC<Props> = ({ navigation }) => {
-  const url = resolveUrlByType(navigation.resultType, navigation.key);
+  const url = resolveUrlBySearchNavigationType(navigation.resultType, navigation.key);
   return (
     <Link href={url} passHref>
       <a className={styles.link}>
@@ -20,28 +20,6 @@ const NavigationItem: React.FC<Props> = ({ navigation }) => {
       </a>
     </Link>
   );
-};
-
-/**
- * Generate the navigation url based on the type.
- *
- * @param {SearchNavigationType} type
- * @param {string | number} key
- * @returns {string}
- */
-const resolveUrlByType = (type: SearchNavigationType, key: string | number): string => {
-  const stringKey = String(key);
-  if (type === SearchNavigationType.AYAH) {
-    return getVerseNavigationUrl(stringKey);
-  }
-  if (type === SearchNavigationType.JUZ) {
-    return `/juz/${key}`;
-  }
-  if (type === SearchNavigationType.PAGE) {
-    return `/page/${key}`;
-  }
-  // for the Surah navigation
-  return `/${key}`;
 };
 
 export default NavigationItem;
